Fix clang-tidy warnings

+++ src/blocks/processes/delay.cpp
@@ -3,7 +3,7 @@
 namespace blocks {
 
 Delay::Delay(float time)
-    : nSamples_(kSampleRate * time), register_(kMaxBufferSize) {}
+    : nSamples_(size_t(kSampleRate * time)), register_(kMaxBufferSize) {}
 
 float Delay::process(float x) {
     register_.push(x);
